requirements.txt
streamlit
pillow
pytesseract
opencv-python-headless
pandas
openpyxl
